Refactor: Add margin bottom in Title component
