+++ src/components/Title/Title.scss
@@ -1,9 +1,13 @@
-.portfolio-title-icon {
+.title-container {
+  margin-bottom: 8px;
+}
+
+.title-icon {
   font-size: 24px;
   margin-right: 8px;
 }
 
-.portfolio-title-text {
+.title-text {
   font-size: 24px;
   font-weight: bold;
 }
